% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ellmer.R
\name{chat_ellmer}
\alias{chat_ellmer}
\title{chat ellmer}
\usage{
chat_ellmer(provider = "ollama")
}
\arguments{
\item{provider}{a provider, such as "ollama", or "claude", or "github"}
}
\description{
a minimal wrapper function to switch which provider is used for each llm_image* function
when ellmer backend is selected, ollamar only supports ollama
}
