% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-prompts.R
\name{edit_prompt}
\alias{edit_prompt}
\title{edit prompt}
\usage{
edit_prompt(prompt)
}
\arguments{
\item{prompt}{a prompt from list_prompts()}
}
\value{
a prompt markdown file to edit
}
\description{
edit a listed prompt installed with kuzco
}
\examples{
\dontrun{
edit_prompt("system-prompt-alt-text.md")
}
}
