% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuzco-package.R
\docType{package}
\name{kuzco-package}
\alias{kuzco}
\alias{kuzco-package}
\title{kuzco: Computer Vision with Large Language Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Make computer vision tasks approachable in R by leveraging Large Language Models. Providing fine-tuned prompts, boilerplate functions, and input/output helpers for common computer vision workflows, such as classifying and describing images. Functions are designed to take images as input and return structured data, helping users build practical applications with minimal code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frankiethull.github.io/kuzco/}
  \item \url{https://github.com/frankiethull/kuzco}
  \item Report bugs at \url{https://github.com/frankiethull/kuzco/issues}
}

}
\author{
\strong{Maintainer}: Frank Hull \email{frankiethull@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Johannes Breuer \email{johannes.breuer@gesis.org} (\href{https://orcid.org/0000-0001-5906-7873}{ORCID}) [contributor]
  \item Jordi Rosell \email{jroselln@gmail.com} (\href{https://orcid.org/0000-0002-4349-1458}{ORCID}) [contributor]
}

}
\keyword{internal}
