% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuzco_app.R
\name{kuzco_app}
\alias{kuzco_app}
\title{shiny kuzco app}
\usage{
kuzco_app()
}
\description{
a simple wrapper of kuzco to make computer vision for everyone.
few-shot via frank hull and shiny assistant
(https://gallery.shinyapps.io/assistant/)
}
