% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-prompts.R
\name{list_prompts}
\alias{list_prompts}
\title{list prompts}
\usage{
list_prompts()
}
\value{
a list of prompts stored within kuzco
}
\description{
list prompts installed with kuzco
}
\examples{
list_prompts()

}
