% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_image_classification.R
\name{llm_image_classification}
\alias{llm_image_classification}
\title{Image Classification using LLMs}
\usage{
llm_image_classification(
  llm_model = "llava-phi3",
  image = system.file("img/test_img.jpg", package = "kuzco"),
  backend = "ellmer",
  ...
)
}
\arguments{
\item{llm_model}{a local LLM model pulled from ollama}

\item{image}{a local image path that has a jpeg, jpg, or png}

\item{backend}{either 'ollamar' or 'ellmer'}

\item{...}{a pass through for other generate args and model args like temperature}
}
\value{
a df with image_classification, primary_object, secondary_object, image_description, image_colors, image_proba_names, image_proba_values
}
\description{
Image Classification using LLMs
}
