% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_image_recognition.R
\name{llm_image_recognition}
\alias{llm_image_recognition}
\title{Image Recognition using LLMs}
\usage{
llm_image_recognition(
  llm_model = "llava-phi3",
  image = system.file("img/test_img.jpg", package = "kuzco"),
  recognize_object = "face",
  backend = "ellmer",
  ...
)
}
\arguments{
\item{llm_model}{a local LLM model pulled from ollama}

\item{image}{a local image path that has a jpeg, jpg, or png}

\item{recognize_object}{an item you want to LLM to look for}

\item{backend}{either 'ollamar' or 'ellmer'}

\item{...}{a pass through for other generate args and model args like temperature. set the temperature to 0 for more deterministic output}
}
\value{
a df with object_recognized, object_count, object_description, object_location
}
\description{
Image Recognition using LLMs
}
