% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_image_sentiment.R
\name{llm_image_sentiment}
\alias{llm_image_sentiment}
\title{Image Sentiment using LLMs}
\usage{
llm_image_sentiment(
  llm_model = "llava-phi3",
  image = "inst/img/test_img.jpg",
  ...
)
}
\arguments{
\item{llm_model}{a local LLM model pulled from ollama}

\item{image}{a local image path that has a jpeg, jpg, or png}

\item{...}{a pass through for other generate args and model args like temperature}
}
\value{
a df with image_sentiment, image_score, sentiment_description, image_keywords
}
\description{
Image Sentiment using LLMs
}
