% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_image.R
\name{view_image}
\alias{view_image}
\title{View Images quickly and easily}
\usage{
view_image(image = system.file("img/test_img.jpg", package = "kuzco"))
}
\arguments{
\item{image}{an image to view}
}
\value{
a plot of the image in a Plots pane
}
\description{
View Images quickly and easily
}
