% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_llm_results.R
\name{view_llm_results}
\alias{view_llm_results}
\title{view llm results as a tidy great table}
\usage{
view_llm_results(llm_results)
}
\arguments{
\item{llm_results}{results from one of the llm_image_* functions}
}
\value{
a great table to view the results neatly
}
\description{
view llm results as a tidy great table
}
