% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuzco-package.R
\docType{package}
\name{kuzco-package}
\alias{kuzco}
\alias{kuzco-package}
\title{kuzco: LLM image classification using ollama in R}
\description{
This package is a designed to use local models for image classification. The prompts and functions are designed to take an input image and supply classification information as an output.
}
\author{
\strong{Maintainer}: Frank Hull \email{frankiethull@gmail.com}

}
\keyword{internal}
