% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuzco-package.R
\docType{package}
\name{kuzco-package}
\alias{kuzco}
\alias{kuzco-package}
\title{kuzco: LLM image classification using ollama in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is a designed to use local models for image classification. The prompts and functions are designed to take an input image and supply classification information as an output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frankiethull.github.io/kuzco/}
}

}
\author{
\strong{Maintainer}: Frank Hull \email{frankiethull@gmail.com}

Other contributors:
\itemize{
  \item Johannes Breuer \email{johannes.breuer@gesis.org} (\href{https://orcid.org/0000-0001-5906-7873}{ORCID}) [contributor]
}

}
\keyword{internal}
