% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_image_extract_text.R
\name{llm_image_extract_text}
\alias{llm_image_extract_text}
\title{Image OCR for Text Extraction using LLMs}
\usage{
llm_image_extract_text(
  llm_model = "llava-phi3",
  image = system.file("img/text_img.jpg", package = "kuzco"),
  backend = "ellmer",
  ...
)
}
\arguments{
\item{llm_model}{a local LLM model pulled from ollama}

\item{image}{a local image path that has a jpeg, jpg, or png}

\item{backend}{either 'ellmer' or 'ollamar'}

\item{...}{a pass through for other generate args and model args like temperature. set the temperature to 0 for more deterministic output}
}
\value{
a df with text
}
\description{
Image OCR for Text Extraction using LLMs
}
